+++ CHANGES.rst
@@ -24,7 +24,7 @@ Changelog
   To handle this, it was necessary to implement a summary serializer for `Meeting`.
   [gbastien]
 - Added `test_restapi_search_items_extra_include_deliberation_images` showing
-  that images are received as base64 data value.
+  that images are received as base64 data value. 
   [gbastien]
 
 1.0b1 (2021-02-03)